requirements.txt
# Core LiveKit Agent
livekit-agents[openai,groq]
livekit-api

# Environment & Configuration  
python-dotenv

# OpenAI SDK (for Groq compatibility)
openai

# FastAPI Backend 
fastapi
uvicorn[standard]

# Additional utilities
pydantic
